+++ crates/miden-lib/src/account/components/mod.rs
@@ -36,6 +36,12 @@ static RPO_FALCON_512_PROCEDURE_ACL_LIBRARY: LazyLock<Library> = LazyLock::new(|
         .expect("Shipped Rpo Falcon 512 Procedure ACL library is well-formed")
 });
 
+// Initialize the NoAuth library only once.
+static NO_AUTH_LIBRARY: LazyLock<Library> = LazyLock::new(|| {
+    let bytes = include_bytes!(concat!(env!("OUT_DIR"), "/assets/account_components/no_auth.masl"));
+    Library::read_from_bytes(bytes).expect("Shipped NoAuth library is well-formed")
+});
+
 /// Returns the Basic Wallet Library.
 pub fn basic_wallet_library() -> Library {
     BASIC_WALLET_LIBRARY.clone()
@@ -54,4 +60,9 @@ pub fn basic_fungible_faucet_library() -> Library {
 /// Returns the Rpo Falcon 512 Procedure ACL Library.
 pub fn rpo_falcon_512_procedure_acl_library() -> Library {
     RPO_FALCON_512_PROCEDURE_ACL_LIBRARY.clone()
+}
+
+/// Returns the NoAuth Library.
+pub fn no_auth_library() -> Library {
+    NO_AUTH_LIBRARY.clone()
 }